RF_01.002.15 | New Item > Folder > test_create_folder_with_same_name_in_different_parent to POM
