+++ tests/test_folder_configuration.py
@@ -1,11 +1,23 @@
 import pytest
 from selenium.webdriver.common.by import By
+from selenium.webdriver.support.wait import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
 
-from test_folder import create_folder
 
 FOLDER_NAME = "TestFolder"
 DISPLAY_NAME = "Display Folder"
 
+def create_folder(driver, name, full_creation=True):
+    driver.find_element(By.XPATH, "//a[contains(@href, '/newJob')]").click()
+    driver.find_element(By.ID, "name").send_keys(name)
+    driver.find_element(By.CLASS_NAME, "com_cloudbees_hudson_plugins_folder_Folder").click()
+    WebDriverWait(driver, 5).until(
+        EC.element_to_be_clickable((By.ID, "ok-button"))
+    ).click()
+    if full_creation:
+        WebDriverWait(driver, 5).until(
+            EC.element_to_be_clickable((By.NAME, "Submit"))
+        ).click()
 
 @pytest.mark.dependency()
 def test_add_display_name_to_folder(browser):
